moved around ODE operations and handlers
