+++ chirho/dynamical/__init__.py
@@ -1,8 +1,1 @@
-import pyro
 
-from chirho.dynamical.ops import simulate
-
-
-class Dynamics(pyro.nn.PyroModule):
-    def forward(self, *args, **kwargs):
-        return simulate(self, *args, **kwargs)
