+++ chirho/dynamical/internals/ODE/__init__.py
@@ -8,10 +8,13 @@ import pyro
 import torch
 import torchdiffeq
 
-from chirho.dynamical.internals.interruption import apply_interruptions, simulate_to_interruption, concatenate
-
-from chirho.dynamical.ODE import ODEDynamics
-from chirho.dynamical.ops import simulate, State, Trajectory
+from chirho.dynamical.internals.interruption import (
+    apply_interruptions,
+    concatenate,
+    simulate_to_interruption,
+)
+from chirho.dynamical.ops import State, Trajectory, simulate
+from chirho.dynamical.ops.ODE import ODEDynamics
 from chirho.indexed.ops import IndexSet, gather, indices_of, union
 from chirho.interventional.handlers import intervene
 from chirho.observational.handlers import condition
