+++ chirho/dynamical/internals/__init__.py
@@ -2,3 +2,4 @@
 # from .indexed import gather, indices_of
 # from .interruption import apply_interruptions, simulate_to_interruption, concatenate
 # from .interventional import intervene
+from .ODE import simulate
