+++ chirho/dynamical/ops/ODE/__init__.py
@@ -9,10 +9,11 @@ import torch
 import torchdiffeq
 
 from chirho.dynamical.internals.interruption import (
+    apply_interruptions,
     concatenate,
+    simulate_to_interruption,
 )
-from chirho.dynamical.internals.interruption import apply_interruptions, simulate_to_interruption
-from chirho.dynamical.ops import Dynamics, simulate, State, Trajectory
+from chirho.dynamical.ops import Dynamics, State, Trajectory, simulate
 from chirho.indexed.ops import IndexSet, gather, indices_of, union
 from chirho.interventional.handlers import intervene
 from chirho.observational.handlers import condition
