+++ tests/dynamical/dynamical_fixtures.py
@@ -3,7 +3,7 @@ import torch
 from pyro.distributions import Normal, Uniform, constraints
 
 from chirho.dynamical.ops import State, Trajectory
-from chirho.dynamical.ODE import ODEDynamics
+from chirho.dynamical.ops.ODE import ODEDynamics
 
 
 class UnifiedFixtureDynamics(ODEDynamics):
