+++ tests/dynamical/test_static_interventions.py
@@ -8,7 +8,7 @@ from chirho.counterfactual.handlers import (
     TwinWorldCounterfactual,
 )
 from chirho.dynamical.handlers import PointIntervention, SimulatorEventLoop
-from chirho.dynamical.ops import simulate, State
+from chirho.dynamical.ops import State, simulate
 from chirho.indexed.ops import IndexSet, gather, indices_of
 from chirho.interventional.ops import intervene
